refactor(ui): split configuration page responsibilities
